src/pyproject.toml
[tool.poetry]
name = "quiz"
version = "0.1.0"
description = ""
authors = ["Hiep <vanhiepdam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-snowflake = "^5.1"
snowflake-id = "^1.0.2"
mysqlclient = "^2.2.4"
redis = "^5.0.8"
python-redis-lock = "^4.0.0"
kafka-python = "^2.0.2"
drf-spectacular = "^0.27.2"
drf-standardized-errors = "^0.14.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.8.0"
flake8 = "^7.1.1"
autoflake = "^2.3.1"
isort = "^5.13.2"
pytest-mock = "^3.14.0"
factory-boy = "^3.3.1"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
